Showing which boost is active with remaining time in inventory.
