+++ Idle_Game/app/src/main/java/com/example/idle_game/ui/views/states/InventoryViewState.kt
@@ -7,6 +7,5 @@ import kotlinx.coroutines.flow.emptyFlow
 
 data class InventoryViewState(
     val shopData: Flow<List<ShopData>> = emptyFlow(),
-    val inventoryData: InventoryData? = null,
-    val selectedItem: ShopData? = null
+    val inventoryData: InventoryData? = null
     )
